+++ CHANGELOG.md
@@ -27,6 +27,8 @@
 - On macOS, file tree icons are now displayed using the theme/system color. On
   all platforms, you can customize the appearance using the selector
   `body .tree-item clr-icon.special` (including platform/mode selectors)
+- Improve the readability mode: Now it will not render a YAML frontmatter and it
+  will additionally treat sentences smarter than before
 
 ## Under the Hood